Move AnalyticsReporter to be next to similar utils
