+++ apps/src/lib/util/AnalyticsReporter.js
